+++ contrib/bc/tests/other.sh
@@ -152,16 +152,19 @@ if [ "$d" = "bc" ]; then
 	printf '5\n0\n' > "$redefine_res"
 
 	"$exe" "$@" --redefine=print -e 'define print(x) { x }' -e 'print(5)' > "$redefine_out"
+	err="$?"
 
 	checktest "$d" "$err" "keyword redefinition" "$redefine_res" "$redefine_out"
 
 	"$exe" "$@" -r "abs" -r "else" -e 'abs = 5;else = 0' -e 'abs;else' > "$redefine_out"
+	err="$?"
 
 	checktest "$d" "$err" "keyword redefinition" "$redefine_res" "$redefine_out"
 
 	if [ "$extra_math" -ne 0 ]; then
 
 		"$exe" "$@" -lr abs -e "perm(5, 1)" -e "0" > "$redefine_out"
+		err="$?"
 
 		checktest "$d" "$err" "keyword not redefined in builtin library" "$redefine_res" "$redefine_out"
 
@@ -176,6 +179,53 @@ if [ "$d" = "bc" ]; then
 	err="$?"
 
 	checkerrtest "$d" "$err" "Keyword redefinition error without BC_REDEFINE_KEYWORDS" "$redefine_out" "$d"
+
+	printf 'pass\n'
+	printf 'Running multiline comment expression file test...'
+
+	multiline_expr_res=""
+	multiline_expr_out="$outputdir/bc_outputs/multiline_expr_results.txt"
+
+	# tests/bc/misc1.txt happens to have a multiline comment in it.
+	"$exe" "$@" -f "$testdir/bc/misc1.txt" > "$multiline_expr_out"
+	err="$?"
+
+	checktest "$d" "$err" "multiline comment in expression file" "$testdir/bc/misc1_results.txt" \
+		"$multiline_expr_out"
+
+	printf 'pass\n'
+	printf 'Running multiline comment expression file error test...'
+
+	"$exe" "$@" -f "$testdir/bc/errors/05.txt" 2> "$multiline_expr_out"
+	err="$?"
+
+	checkerrtest "$d" "$err" "multiline comment in expression file error" \
+		"$multiline_expr_out" "$d"
+
+	printf 'pass\n'
+	printf 'Running multiline string expression file test...'
+
+	# tests/bc/strings.txt happens to have a multiline string in it.
+	"$exe" "$@" -f "$testdir/bc/strings.txt" > "$multiline_expr_out"
+	err="$?"
+
+	checktest "$d" "$err" "multiline string in expression file" "$testdir/bc/strings_results.txt" \
+		"$multiline_expr_out"
+
+	printf 'pass\n'
+	printf 'Running multiline string expression file error test...'
+
+	"$exe" "$@" -f "$testdir/bc/errors/16.txt" 2> "$multiline_expr_out"
+	err="$?"
+
+	checkerrtest "$d" "$err" "multiline string in expression file with backslash error" \
+		"$multiline_expr_out" "$d"
+
+	"$exe" "$@" -f "$testdir/bc/errors/04.txt" 2> "$multiline_expr_out"
+	err="$?"
+
+	checkerrtest "$d" "$err" "multiline string in expression file error" \
+		"$multiline_expr_out" "$d"
 
 	printf 'pass\n'