+++ contrib/bc/NEWS.md
@@ -1,5 +1,10 @@
 # News
 
+## 5.2.3
+
+This is a production release that fixes one bug, a parse error when passing a
+file to `bc` using `-f` if that file had a multiline comment or string in it.
+
 ## 5.2.2
 
 This is a production release that fixes one bug, a segmentation fault if